requirements.txt
boto3~=1.36.4
botocore~=1.36.3
certifi
charset-normalizer
idna
jmespath
psycopg2-binary~=2.9.10
python-dateutil
requests~=2.32.3
s3transfer
six
urllib3
pandas~=2.2.3
pydantic~=2.10.5
fsspec
s3fs
pyarrow~=19.0.0
fastparquet
snowflake-connector-python
pyodbc~=5.2.0
cryptography~=44.0.0
PyJWT~=2.10.1